feat: puyo split test coverage, improved naming, bug fix
Couple issues - we now skip the bottom row completely in this check (to
avoid having to care if 'below' the bottom is empty/off-the-grid). we
also correct the `if p_below: continue` logic, which was backwards - we
skip when there is already a piece below us.

Also improves some to_pretty() to help with debugging.