refactor: pull bucket tick into grid step logic
Makes more sense for the grid to own this step(opts) function - it's
important to get the order right, and it'll help the consumers handle
various game modes correctly.

the bucket/grid tests are a bit duped right now, but it's nice to have
both 'tick' and 'step' unit tested at least, so i'm leaving the bucket
tests for now.