feat: ticking and effects in lock step!
Finally gets a handle on the flow - waiting for animations to complete
before going to the next step. Huzzah!

A bit of a mess, and difficult to test - should refactor to separate
determining the next step() action separate from performing it.

the action_queue was a nice fix here, and a few things can be cleaned
up (e.g. the extra set of on_blah_complete signals).