fix: prevent multiple pieces from adding at once
User input was causing do_next_action to fire multiple times, which in
some cases led to multiple pieces starting at once. This adds a new
feature - add_piece_next_tick() - to opt-in to a new piece being added
on the next tick. We should still prevent next_action from firing
multiple times, but that's next and this is more resilient for all actions.